refactor: implement feature sliced disign
