--- src/App.tsx
@@ -1,16 +1,0 @@
-import { useNavigate } from "react-router-dom";
-import "./App.css";
-import Dashboard from "./rootes/dashboard";
-import Root from "./rootes/dashboard";
-
-function App() {
-  
-  return (
-    <div>
-      <h1>Choose the language</h1>
-      {/* <Dashboard /> */}
-    </div>
-  );
-}
-
-export default App;
