+++ src/app/App.tsx
@@ -1,0 +1,14 @@
+import { useNavigate } from "react-router-dom";
+
+import Dashboard from "./routes/dashboard";
+
+function App() {
+  return (
+    <div>
+      {/* <h1>Choose the language</h1> */}
+      <Dashboard />
+    </div>
+  );
+}
+
+export default App;
