+++ src/components/Apps/react/counter/fix/Counter.js
@@ -1,11 +1,7 @@
 import React, { useState } from "react";
+//Counter should begin from 0 and increase and decrease by 1
 
 const Counter = () => {
-  const style = {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-  };
   const [counter, setCounter] = useState(1);
 
   const increment = () => {
@@ -16,13 +12,21 @@ const Counter = () => {
     setCounter(counter - 2);
   };
   return (
-    <div style={style}>
-      <h1 className="text-xl font bold">
-        Counter should begin from 0 and increase and decrease by 1
-      </h1>
-      <button onClick={decrement}>-</button>
-      <h2 style={{ margin: "0 10px" }}>counter: {counter}</h2>
-      <button onClick={increment}>+</button>
+    <div className="flex items-center justify-center h-96 bg-gray-100">
+  
+      <button
+        onClick={decrement}
+        className="bg-red-500 text-white py-2 px-4 rounded hover:bg-red-600"
+      >
+        -
+      </button>
+      <h2 className="mx-4">Counter: {counter}</h2>
+      <button
+        onClick={increment}
+        className="bg-green-500 text-white py-2 px-4 rounded hover:bg-green-600"
+      >
+        +
+      </button>
     </div>
   );
 };
